refactor : 로또 번호 선언 정적 메서드 삭제. 테스트 클래스의 @DisplayName 삽입
