+++ src/main/java/lotto/view/LottoOutputView.java
@@ -1,6 +1,8 @@
 package lotto.view;
 
 import lotto.domain.*;
+import lotto.dto.BuyingListDto;
+import lotto.dto.LottoDto;
 
 import java.util.List;
